Finish most base skeleton of trip section
