+++ client/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@lottiefiles/react-lottie-player": "^3.5.3",
     "@mui/icons-material": "^5.14.9",
     "@mui/material": "^5.14.13",
     "@mui/styles": "^5.14.10",
@@ -32,6 +33,7 @@
     "react": "18.2.0",
     "react-date-range": "^1.4.0",
     "react-day-picker": "^8.9.1",
+    "react-device-frameset": "^1.3.2",
     "react-dom": "18.2.0",
     "react-i18next": "^13.3.1",
     "react-swipeable-views": "^0.14.0",
