+++ client/src/app/[locale]/(components)/(home)/GridCards.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Box, Typography } from "@mui/material";
-import TripObject from "../../(object-types)/TripObject";
+import TripObject from "../(object-types)/TripObject";
 import SpringNote from "../(notes)/SpringNote";
 
 interface GridCardProps {
