+++ client/src/app/[locale]/(components)/(object-types)/TripObject.tsx
